# openapi.yaml
openapi: 3.1.0
info:
  title: IoT and IM Microservices API
  description: |
    Formal API specification matching the provided Python microservices:
    - **IMTelegram Service**: Port 5000 (Prefix: /IMTelegram)
    - **IMLine Service**: Port 5001 (Prefix: /IMLine)
    - **ESP32 Service**: Port 5002/5010 (Prefix: /ESP32)
    - **Raspberry Pi Service**: Port 5003/5011 (Prefix: /Pi)
  version: 2.0.2
  contact:
    name: Rocer52
    email: batte9306@gmail.com
servers:
  - url: http://localhost:5000
    description: IM Telegram Service
  - url: http://localhost:5001
    description: IM Line Service
  - url: http://localhost:5002
    description: ESP32 Service (IoT)
  - url: http://localhost:5003
    description: Raspberry Pi Service (IoT)
...
paths:
...
  /Pi/{device_id}/Enable:
    get:
      tags:
        - RaspberryPi
      summary: Enable Pi Device (GET)
      operationId: enablePiGet
      parameters:
        - $ref: '#/components/parameters/path_device_id'
        - $ref: '#/components/parameters/chat_id'
        - $ref: '#/components/parameters/timestamp'
        - $ref: '#/components/parameters/signature'
        - $ref: '#/components/parameters/username'
        - $ref: '#/components/parameters/bot_token'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
    post:
      tags:
        - RaspberryPi
      summary: Enable Pi Device (POST)
      operationId: enablePiPost
      parameters:
        - $ref: '#/components/parameters/path_device_id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceActionBody'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
  /Pi/{device_id}/Disable:
    get:
      tags:
        - RaspberryPi
      summary: Disable Pi Device (GET)
      operationId: disablePiGet
      parameters:
        - $ref: '#/components/parameters/path_device_id'
        - $ref: '#/components/parameters/chat_id'
        - $ref: '#/components/parameters/timestamp'
        - $ref: '#/components/parameters/signature'
        - $ref: '#/components/parameters/username'
        - $ref: '#/components/parameters/bot_token'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
    post:
      tags:
        - RaspberryPi
      summary: Disable Pi Device (POST)
      operationId: disablePiPost
      parameters:
        - $ref: '#/components/parameters/path_device_id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceActionBody'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
  /Pi/{device_id}/GetStatus:
    get:
      tags:
        - RaspberryPi
      summary: Get Pi Status (GET)
      operationId: getPiStatusGet
      parameters:
        - $ref: '#/components/parameters/path_device_id'
        - $ref: '#/components/parameters/chat_id'
        - $ref: '#/components/parameters/timestamp'
        - $ref: '#/components/parameters/signature'
        - $ref: '#/components/parameters/username'
        - $ref: '#/components/parameters/bot_token'
      responses:
        '200':
          description: Status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
    post:
      tags:
        - RaspberryPi
      summary: Get Pi Status (POST)
      operationId: getPiStatusPost
      parameters:
        - $ref: '#/components/parameters/path_device_id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceActionBody'
      responses:
        '200':
          description: Status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
  /signature:
    post:
      tags:
        - Utilities
      summary: Verify or Generate Signature
      description: Available on both ESP32 and Pi services.
      operationId: manageSignature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignatureRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignatureResponse'
        '403':
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
          default: false
        message:
          type: string
          example: "Invalid request"
      required:
        - ok
        - message
    DeviceResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Device enabled"
        state:
          type: string
          example: "on"
        device_id:
          type: string
          example: "esp32_light_001"
        username:
          type: string
          example: "user123"
      required:
        - status
        - message
        - state
        - device_id
    DeviceActionBody:
      type: object
      description: JSON body for POST requests to IoT devices
      properties:
        chat_id:
          type: string
          example: "123456789"
        timestamp:
          type: string
          example: "1700000000"
        signature:
          type: string
          example: "base64_encoded_signature_example"
        username:
          type: string
          example: "user123"
        bot_token:
          type: string
          example: "123456789:ABCDEF_ExampleBotToken"
      required:
        - chat_id
        - timestamp
        - signature
    SignatureRequest:
      type: object
      properties:
        chat_id:
          type: string
          example: "123456789"
        timestamp:
          type: string
          example: "1700000000"
        signature:
          type: string
          example: "base64_encoded_signature_example"
      required:
        - chat_id
        - timestamp
        - signature
    SignatureResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Signature valid"
        chat_id:
          type: string
          example: "123456789"
        timestamp:
          type: string
          example: "1700000000"
        signature:
          type: string
          example: "base64_encoded_signature_example"
        username:
          type: string
          example: "user123"
        bot_token:
          type: string
          example: "123456789:ABCDEF_ExampleBotToken"
      required:
        - status
        - message
...
  parameters:
...
    path_device_id:
      name: device_id
      in: path
      description: Device ID in the URL path
      required: true
      schema:
        type: string
        enum:
          - esp32_light_001
          - esp32_fan_002
          - raspberrypi_light_001
          - raspberrypi_fan_002
        example: "esp32_light_001"
    chat_id:
      name: chat_id
      in: query
      description: Chat ID
      required: false
      schema:
        type: string
        example: "123456789"
...
    bot_token:
      name: bot_token
      in: query
      description: Bot token
      required: false
      schema:
        type: string
        example: "123456789:ABCDEF_ExampleBotToken"
...
    username:
      name: username
      in: query
      description: User display name
      required: false
      schema:
        type: string
        example: "user123"
    timestamp:
      name: timestamp
      in: query
      description: Security timestamp
      required: false
      schema:
        type: string
        example: "1700000000"
    signature:
      name: signature
      in: query
      description: Security signature
      required: false
      schema:
        type: string
        example: "base64_encoded_signature_example"